Refactor♻️: 장소정보카드 헤더로 리팩토링
